+++ api/nblib/forcecalculator.h
@@ -1,7 +1,7 @@
 /*
  * This file is part of the GROMACS molecular simulation package.
  *
- * Copyright (c) 2020, by the GROMACS development team, led by
+ * Copyright (c) 2020,2021, by the GROMACS development team, led by
  * Mark Abraham, David van der Spoel, Berk Hess, and Erik Lindahl,
  * and including many others, as listed in the AUTHORS file in the
  * top-level source directory and at http://www.gromacs.org.
@@ -99,9 +99,9 @@ public:
      * \param coordinates The coordinates to be placed on grids
      * \param[in] box The system simulation box
      */
-    void updatePairList(gmx::ArrayRef<const int> particleInfoAllVdW,
-                        gmx::ArrayRef<Vec3>      coordinates,
-                        const Box&               box);
+    void updatePairList(gmx::ArrayRef<const int64_t> particleInfoAllVdW,
+                        gmx::ArrayRef<Vec3>          coordinates,
+                        const Box&                   box);
 
 private:
     //! GROMACS force calculator to compute forces
