+++ src/gromacs/mdlib/forcerec.cpp
@@ -288,7 +288,7 @@ makeAtomInfoForEachMoleculeBlock(const gmx_mtop_t& mtop, const t_forcerec* fr)
             for (int a = 0; a < molt.atoms.nr; a++)
             {
                 const t_atom& atom = molt.atoms.atom[a];
-                int& atomInfo      = atomInfoOfMoleculeBlock.atomInfo[moleculeOffsetInBlock + a];
+                int64_t& atomInfo  = atomInfoOfMoleculeBlock.atomInfo[moleculeOffsetInBlock + a];
 
                 /* Store the energy group in atomInfo */
                 int gid  = getGroupType(mtop.groups,
@@ -345,12 +345,12 @@ makeAtomInfoForEachMoleculeBlock(const gmx_mtop_t& mtop, const t_forcerec* fr)
     return atomInfoForEachMoleculeBlock;
 }
 
-static std::vector<int> expandAtomInfo(const int                                             nmb,
-                                       gmx::ArrayRef<const gmx::AtomInfoWithinMoleculeBlock> atomInfoForEachMoleculeBlock)
+static std::vector<int64_t> expandAtomInfo(const int nmb,
+                                           gmx::ArrayRef<const gmx::AtomInfoWithinMoleculeBlock> atomInfoForEachMoleculeBlock)
 {
     const int numAtoms = atomInfoForEachMoleculeBlock[nmb - 1].indexOfLastAtomInMoleculeBlock;
 
-    std::vector<int> atomInfo(numAtoms);
+    std::vector<int64_t> atomInfo(numAtoms);
 
     int mb = 0;
     for (int a = 0; a < numAtoms; a++)
