+++ src/gromacs/mdtypes/atominfo.h
@@ -49,7 +49,7 @@
 namespace gmx
 {
 
-/*! \brief Constants whose bit describes properties of atoms in
+/*! \brief Constants whose bit describes a property of an atom in
  * AtomInfoWithinMoleculeBlock.atomInfo.
  *
  * No bit should exceed 1 << 63, so that it fits into a 64-bit
@@ -59,16 +59,16 @@ namespace gmx
  * here, reserving bits 0-7 for the energy-group ID.
  */
 //! \{
-static constexpr int sc_atomInfo_FreeEnergyPerturbation = 1 << 15;
-static constexpr int sc_atomInfo_Exclusion              = 1 << 17;
-static constexpr int sc_atomInfo_Constraint             = 1 << 20;
-static constexpr int sc_atomInfo_Settle                 = 1 << 21;
-static constexpr int sc_atomInfo_BondCommunication      = 1 << 22;
-static constexpr int sc_atomInfo_HasVdw                 = 1 << 23;
-static constexpr int sc_atomInfo_HasCharge              = 1 << 24;
+static constexpr int64_t sc_atomInfo_FreeEnergyPerturbation = 1 << 15;
+static constexpr int64_t sc_atomInfo_Exclusion              = 1 << 17;
+static constexpr int64_t sc_atomInfo_Constraint             = 1 << 20;
+static constexpr int64_t sc_atomInfo_Settle                 = 1 << 21;
+static constexpr int64_t sc_atomInfo_BondCommunication      = 1 << 22;
+static constexpr int64_t sc_atomInfo_HasVdw                 = 1 << 23;
+static constexpr int64_t sc_atomInfo_HasCharge              = 1 << 24;
 //! \}
 //! The first 8 bits are reserved for energy-group ID
-static constexpr int sc_atomInfo_EnergyGroupIdMask = 0b11111111;
+static constexpr int64_t sc_atomInfo_EnergyGroupIdMask = 0b11111111;
 
 /*! \brief Contains information about each atom in a molecule block of
  * the global topology. */
@@ -94,7 +94,7 @@ struct AtomInfoWithinMoleculeBlock
      *
      * The vector needs to be indexed accordingly.
      */
-    std::vector<int> atomInfo;
+    std::vector<int64_t> atomInfo;
 };
 
 } // namespace gmx
